tensor experiments

+++ MiniFlow/Common.h
@@ -4,7 +4,7 @@ namespace miniflow
 {
 	// Typedefs:
 	typedef double Scalar;
-	typedef size_t Index;
+	typedef unsigned Index;
 
 	// math constants
 	Scalar EXP = 2.71828182845904523536;
